Added bsort.  It works, although the actual memory usage definitely exceeds the amount of memory I allocate.

+++ butil/bsort.h
@@ -8,7 +8,7 @@ typedef struct {
 	int32_t startPos;
 	int32_t endContig;
 	int32_t endPos;
-	int32_t memory; /* in MB */
+	int64_t memory; 
 	int64_t numEntries;
 	FILE *FP;
 	char *FileName;
